backend/requirements.txt
Django==4.0.8
djangorestframework==3.13.1
djoser
drf-yasg
django-cors-headers
django-filter==22.1
FileBrowser==1.1.3
django-filebrowser-no-grappelli==4.0.1
django-tinymce==3.5.0
django-baton
django-extra-settings==0.6.1
django-phonenumber-field[phonenumbers]==7.0.0
django-colorfield
django_crispy_forms==1.14.0
django-watson
crispy-bootstrap5
celery[redis]==5.2.7
django-celery-results==2.4.0
django-celery-beat==2.3.0
django-allauth
django-debug-toolbar
redis==4.3.4
pyQiwiP2P==2.0.6
shortuuid==1.0.9
Markdown==3.4.1
psycopg2-binary==2.8.6
Pillow==9.2.0
gunicorn==20.1.0
art==5.7
pytz
tzdata
